add fixes

+++ migrations/sql/create-trackings-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS trackings(
     chat_id BIGINT,
-    link_id INTEGER,
+    link_id BIGINT,
     FOREIGN KEY (chat_id) REFERENCES chats (chat_id) ON DELETE CASCADE,
     FOREIGN KEY (link_id) REFERENCES links (link_id) ON DELETE CASCADE,
     PRIMARY KEY (link_id, chat_id)
